requirements_2.txt
peft
diffusers
torch
torchvision
Pillow